Core Customer Experience/Commerce/CommerceOpenAISample/requirements.txt
requests
tenacity
streamlit
pyodbc
openai==1.2.4
python-dotenv
pandas
azure-identity